requirements.txt
PyQt5==5.15.9
PyPDF2==3.0.1
reportlab==4.0.4
Pillow==10.0.0